Update patch set 1
Patch Set 1: Code-Review+1

This is acceptable. Though to be pedantic it should say Main profile (0) and High profile (1) because Main and High are the defined names in Annex A. That said, it's probably defined elsewhere in the spec as Profile 0 and Profile 1. ;)  LGTM

Patch-set: 1
Reviewer: Gerrit User 1135519 <1135519@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1