Do not copy past the end of the buffer in ExtendFrame().
There is a latent bug in ExtendFrame() -- copying |stride| pixels to the
bottommost border of the frame buffer may shoot past the end of the buffer.
Fix this bug by copying up to width + right pixels past the beginnig of
the bottommost border.

This latent bug will be manifested by passing a smaller |left| argument
to ExtendFrame().

PiperOrigin-RevId: 277593753
Change-Id: Idb53cb783ab3814d43d6f8ebb7f627ea24e523ca