threadpool: allow ..._USE_STD_MUTEX to be set externally
LIBGAV1_THREADPOOL_USE_STD_MUTEX behaves the same as previously if
unset: choosing a value appropriate for the detected platform.

note this could be made an option in cmake to better control the absl
dependencies; this change only updates the #if structure in
threadpool.h for use with other build systems.

BUG=b/142835450
PiperOrigin-RevId: 279394836
Change-Id: I7fcbff3e578899c770860eeee1ac5d56ced34bf3

Change-Id: I19add02cf01798207e5e822dcbba224bbdef02b6