Add new kStatusNothingToDequeue status code
This status code will be returned only from DequeueFrame(). This
will be returned only in the following case: DequeueFrame is called
when there are no enqueued frames that are being processed (i.e.)
the input frame queue is empty.

This is useful in distinguishing the above mentioned case from the
case where there are no displayable frames in the temporal unit.

Behavior before:
Both cases: return value is kStatusOk and out_ptr is set to nullptr.

Behavior After:
No enqueued frames available: return value is kStatusNothingToDequeue
No displayable frame in the enqueued frame: return value is kStatusOk
with out_ptr set to nullptr.

PiperOrigin-RevId: 312310874
Change-Id: I80e0df69d4780cea2df603a8a0dd721ce7f898a8