Implement output_all_layers for frame parallel mode
Add support for output_all_layers in frame parallel mode. One of the
key differences is that the layers may finish decoding out of order.
We would have to return them in the order in which they appeared in the
input (otherwise we will lose the deterministic nature of the decoder
output). So we need to remember the input order and sort the output using
that ordering before returning the frames to the application.

PiperOrigin-RevId: 313277088
Change-Id: I3706a96adf85d5e171c5d10dfe4c6f5635261af4