FileReader::Open: explicitly move return value
This function returns a std::unique_ptr<FileReaderInterface> which
FileReader derives from. In C++11 overload resolution was not considered
for local variables:

12.8 Copying and moving class objects (32)
When the criteria for elision of a copy operation are met or would be
met save for the fact that the source object is a function parameter,
and the object to be copied is designated by an lvalue, overload
resolution to select the constructor for the copy is first performed as
if the object were designated by an rvalue.

https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2012/n3337.pdf

Some compilers may have adopted C++14 behavior so this change may not
always be strictly necessary:

When the criteria for elision of a copy/move operation are met, but not
for an exception-declaration, and the object to be copied is designated
by an lvalue, or when the expression in a return statement is a
(possibly parenthesized) id-expression that names an object with
automatic storage duration declared in the body or
parameter-declaration-clause of the innermost enclosing function or
lambda-expression, overload resolution to select the constructor for the
copy is first performed as if the object were designated by an rvalue

https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2014/n4296.pdf

This fixes compilation with some older versions of clang (observed with
3.8).

Bug: b/247299694
PiperOrigin-RevId: 477575208
Change-Id: I15989ea73143134d70d82618763ccd1da0c49742