add common_dsp_test.cc
Use this to call AVX2/SSE4 specific functions, formerly in the
common_{avx2,sse4}_test tests. This avoids invoking the tests without
checking for CPU support.

The test fixture needs to be in a separate file to avoid AVX2/SSE4
instructions being generated in the fixture or gtest support code.
common_neon_test.cc is left as is since NEON is a requirement for Arm
builds currently.

Bug: b/289624514
PiperOrigin-RevId: 550705795
Change-Id: Idd984e3f94d0c5b1e9d1bbc8d2fc376e60c17cc5